feat: 입고예정, 상세 검색 기능 구현
