+++ wms/src/main/java/com/shinlogis/wms/AppMain.java
@@ -28,8 +28,8 @@ import com.shinlogis.wms.common.config.Config;
 import com.shinlogis.wms.common.config.Page;
 import com.shinlogis.wms.common.util.DBManager;
 import com.shinlogis.wms.headquarters.model.HeadquartersUser;
-import com.shinlogis.wms.inOutBound.view.InboundDetailPage;
-import com.shinlogis.wms.inOutBound.view.InboundReceiptPage;
+import com.shinlogis.wms.inbound.view.InboundDetailPage;
+import com.shinlogis.wms.inbound.view.InboundReceiptPage;
 import com.shinlogis.wms.inventory.view.InventoryPage;
 import com.shinlogis.wms.location.model.LocationUser;
 import com.shinlogis.wms.main.view.MainPage;
