+++ wms/src/main/java/com/shinlogis/wms/inbound/model/IODetail.java
@@ -1,4 +1,4 @@
-package com.shinlogis.wms.inOutBound.model;
+package com.shinlogis.wms.inbound.model;
 
 import java.sql.Date;
