+++ wms/src/main/java/com/shinlogis/wms/inbound/model/IOReceipt.java
@@ -1,6 +1,7 @@
-package com.shinlogis.wms.inOutBound.model;
+package com.shinlogis.wms.inbound.model;
 
-import java.util.Date;
+
+import java.sql.Date;
 
 import com.shinlogis.wms.headquarters.model.HeadquartersUser;
 import com.shinlogis.wms.location.model.Location;
